remove unused enum
